requirements.txt
arrow==1.2.3
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
comtypes==1.2.0
cryptography==41.0.7
dnspython==2.2.1
emoji==2.2.0
idna==3.4
lxml==4.9.3
MouseInfo==0.1.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==9.3.0
playsound==1.3.0
plum-dispatch==1.7.4
PyAutoGUI==0.9.53
pycparser==2.21
PyDirectInput==1.0.4
pygame==2.1.3.dev8
PyGetWindow==0.0.9
pymongo==4.3.3
PyMsgBox==1.0.9
pynput==1.7.6
PyPDF2==3.0.1
pypdfium2==4.25.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-docx==1.1.0
python-pptx==0.6.23
pytweening==1.0.4
pywin32==306
reportlab==4.0.8
requests==2.28.1
six==1.16.0
Spire.Presentation==8.9.0
typing_extensions==4.9.0
urllib3==1.26.13
XlsxWriter==3.1.9
